Test refactoring and cleanup
- Clean up all of mocha CLI arguments (they were redundant)
- Add a single test entrypoint (setup.js) where we load all the global dependencies
- Remove all require('expect) from all files (they were not needed now)
- Remove all dotEnv require (not needed anymore, initiated through setup.js)
- Organize remaining tests to integration/unit folders respectfuly.
- Separated some tests into their own files
