+++ package.json
@@ -56,7 +56,7 @@
   "types": "types",
   "scripts": {
     "test": "tools/scripts/test.sh",
-    "test:unit" : "mocha --ui bdd -R spec --recursive test/unit",
+    "test:unit" : "mocha --recursive test/unit",
     "test-with-temp-cloud": "tools/scripts/tests-with-temp-cloud.sh",
     "dtslint": "tools/scripts/ditslint.sh",
     "lint": "tools/scripts/lint.sh",
