+++ test/integration/api/search/search_spec.js
@@ -1,14 +1,6 @@
-require('dotenv').load({
-  silent: true
-});
-
-const expect = require('expect.js');
 const Q = require('q');
 const cloudinary = require('../../../../cloudinary');
 const helper = require("../../../spechelper");
-
-
-
 const testConstants = require('../../../testUtils/testConstants');
 
 const {
