+++ test/integration/cache_spec.js
@@ -1,18 +1,18 @@
 var options;
 const { describe, before, it } = require('mocha');
-const expect = require("expect.js");
-const path = require('path');
-const helper = require("./spechelper");
-const cloudinary = require('../cloudinary').v2;
 
-const FileKeyValueStorage = require(`../${helper.libPath}/cache/FileKeyValueStorage`);
-const KeyValueCacheAdapter = require(`../${helper.libPath}/cache/KeyValueCacheAdapter`);
+const path = require('path');
+const helper = require("../spechelper");
+const cloudinary = require('../../cloudinary').v2;
+
+const FileKeyValueStorage = require(`../../${helper.libPath}/cache/FileKeyValueStorage`);
+const KeyValueCacheAdapter = require(`../../${helper.libPath}/cache/KeyValueCacheAdapter`);
 
 const Cache = cloudinary.Cache;
 const IMAGE_FILE = helper.IMAGE_FILE;
 const PUBLIC_ID = "dummy";
 const BREAKPOINTS = [5, 3, 7, 5];
-const testConstants = require('./testUtils/testConstants');
+const testConstants = require('../testUtils/testConstants');
 
 const {
   TIMEOUT,
