+++ test/mocha.opts
@@ -1,3 +1,3 @@
---require expect.js
+--require './test/setup.js'
 -R spec
 --ui bdd
