+++ test/setup.js
@@ -1,0 +1,9 @@
+global.expect = require('expect.js');
+require('./testUtils/testBootstrap');
+
+
+require('dotenv').load({
+  silent: true
+});
+
+
