+++ test/testUtils/assertions/produceUrl.js
@@ -1,4 +1,3 @@
-const expect = require('expect.js');
 const cloneDeep = require('lodash/cloneDeep');
 const cloudinary = require('../../../cloudinary');
