+++ test/unit/cache/FileKeyValueCache_spec.js
@@ -1,4 +1,3 @@
-const expect = require("expect.js");
 const fs = require('fs');
 const path = require('path');
 const helper = require("../../spechelper");
