+++ test/unit/cloudinary_spec.js
@@ -1,6 +1,5 @@
-const expect = require("expect.js");
-const cloudinary = require("../cloudinary.js");
-const createTestConfig = require('./testUtils/createTestConfig');
+const cloudinary = require("../../cloudinary.js");
+const createTestConfig = require('../testUtils/createTestConfig');
 
 describe("cloudinary", function () {
   beforeEach(function () {
@@ -792,93 +791,5 @@ describe("cloudinary", function () {
     };
     result = cloudinary.utils.url("sample.jpg", options);
     expect(result).to.eql('http://res.cloudinary.com/test123/image/upload/s--v2fTPYTu--/sample.jpg');
-  });
-  describe("getUserAgent", function () {
-    var platform = "";
-    before(function () {
-      platform = cloudinary.utils.userPlatform;
-      cloudinary.utils.userPlatform = "";
-    });
-    after(function () {
-      cloudinary.utils.userPlatform = platform;
-    });
-    it("should add a user platform to USER_AGENT", function () {
-      cloudinary.utils.userPlatform = "Spec/1.0 (Test)";
-      expect(cloudinary.utils.getUserAgent()).to.match(/Spec\/1.0 \(Test\) CloudinaryNodeJS\/[\d.]+/);
-    });
-  });
-  describe("config", function () {
-    let cloudinaryUrlBackup;
-    let accountUrlBackup;
-    before(function () {
-      cloudinaryUrlBackup = process.env.CLOUDINARY_URL;
-      accountUrlBackup = process.env.CLOUDINARY_ACCOUNT_URL;
-    });
-    describe("CLOUDINARY_URL", function () {
-      after(function () {
-        process.env.CLOUDINARY_URL = cloudinaryUrlBackup || '';
-        cloudinary.config(true);
-      });
-      it("should allow nested values in CLOUDINARY_URL", function () {
-        process.env.CLOUDINARY_URL = "cloudinary://key:secret@test123?foo[bar]=value";
-        cloudinary.config(true);
-        const foo = cloudinary.config().foo;
-        expect(foo && foo.bar).to.eql('value');
-      });
-      it("should load a properly formatted CLOUDINARY_URL", function () {
-        process.env.CLOUDINARY_URL = "cloudinary://123456789012345:ALKJdjklLJAjhkKJ45hBK92baj3@test";
-        cloudinary.config(true);
-      });
-      it("should not be sensitive to case in CLOUDINARY_URL's protocol", function () {
-        process.env.CLOUDINARY_URL = "CLouDiNaRY://123456789012345:ALKJdjklLJAjhkKJ45hBK92baj3@test";
-        cloudinary.config(true);
-      });
-      it("should throw error when CLOUDINARY_URL doesn't start with 'cloudinary://'", function () {
-        process.env.CLOUDINARY_URL = "https://123456789012345:ALKJdjklLJAjhkKJ45hBK92baj3@test?cloudinary=foo";
-        try {
-          cloudinary.config(true);
-          expect().fail();
-        } catch (err) {
-          expect(err.message).to.eql("Invalid CLOUDINARY_URL protocol. URL should begin with 'cloudinary://'");
-        }
-      });
-      it("should not throw an error when CLOUDINARY_URL environment variable is missing", function () {
-        delete process.env.CLOUDINARY_URL;
-        cloudinary.config(true);
-      });
-    });
-    describe("CLOUDINARY_ACCOUNT_URL", function () {
-      after(function () {
-        process.env.CLOUDINARY_ACCOUNT_URL = accountUrlBackup || '';
-        cloudinary.config(true);
-      });
-      it("should allow nested values in CLOUDINARY_ACCOUNT_URL", function () {
-        process.env.CLOUDINARY_ACCOUNT_URL = "account://key:secret@test123?foo[bar]=value";
-        cloudinary.config(true);
-        const foo = cloudinary.config().foo;
-        expect(foo && foo.bar).to.eql('value');
-      });
-      it("should load a properly formatted CLOUDINARY_ACCOUNT_URL", function () {
-        process.env.CLOUDINARY_ACCOUNT_URL = "account://635412789012345:ALKJdjklLJAjhkKJ45hBK92tam2@test1";
-        cloudinary.config(true);
-      });
-      it("should not be sensitive to case in CLOUDINARY_ACCOUNT_URL's protocol", function () {
-        process.env.CLOUDINARY_ACCOUNT_URL = "aCCouNT://635283989012345:ALKGssklLJAjhkKJ45hBK92tas5@test1";
-        cloudinary.config(true);
-      });
-      it("should throw error when CLOUDINARY_ACCOUNT_URL doesn't start with 'account://'", function () {
-        process.env.CLOUDINARY_ACCOUNT_URL = "https://635283989012345:ALKGssklLJAjhkKJ45hBK92tas5@test1?account=foo";
-        try {
-          cloudinary.config(true);
-          expect().fail();
-        } catch (err) {
-          expect(err.message).to.eql("Invalid CLOUDINARY_ACCOUNT_URL protocol. URL should begin with 'account://'");
-        }
-      });
-      it("should not throw an error when CLOUDINARY_ACCOUNT_URL environment variable is missing", function () {
-        delete process.env.CLOUDINARY_ACCOUNT_URL;
-        cloudinary.config(true);
-      });
-    });
   });
 });
