+++ test/utils/get_sdk_versionID_spec.js
@@ -1,5 +1,4 @@
-let path = require('path');
-const expect = require("expect.js");
+const path = require('path');
 const mock = require('mock-fs');
 const getSDKVersionID = require('../../lib/utils/encoding/sdkVersionID/getSDKVersionID');
