+++ test/utils/utils_spec.js
@@ -1,8 +1,3 @@
-require('dotenv').load({
-  silent: true
-});
-
-const expect = require("expect.js");
 const fs = require('fs');
 const os = require('os');
 const defaults = require('lodash/defaults');
